+++ mapboard/topology_manager/commands/__init__.py
@@ -1,5 +1,6 @@
 from .clean_topology import clean_topology
 from .create_tables import create_tables
+from .edge_relations import rebuild_edge_relations, validate_edge_relations
 from .update_contacts import update_contacts
 from .update_faces import update_faces
 from .update_topology import update